requirements.txt
pyyaml
celery
redis
flower
